+++ frontend/js/components/battle-details/PokemonCard.js
