+++ frontend/js/pages/SettledBattlesList.js
@@ -3,8 +3,8 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import { getSettledBattlesList } from '../actions/battles-list';
-import Loading from '../components/loading';
-import PageTitle from '../components/title';
+import Loading from '../components/Loading';
+import PageTitle from '../components/Title';
 
 function SettledBattleItem({ battle }) {
   return (
